backend/requirements.txt
asgiref==3.9.1
dataclasses==0.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djongo5==1.3.9
dnspython==2.7.0
pymongo==4.13.2
pyspark==3.5.5
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
